Create learning endpoint (#2)
Add a function to fetch those day programming newsletters we all signup for and post them to slack channels.

This new function will do the fetch on invocation and concurrently query each RSS feed for the latest post. If the post has not been posted (a record is not in the DB) then it will post to the correct channel. After a successful post a new database record will be written to make sure we don't share the post multiple times.